Added CompletenessQueryBuilder
