+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -83,5 +83,4 @@ services:
         class: %pim_catalog.manager.completeness.class%
         arguments:
             - '@doctrine'
-            - %pim_catalog.entity.product.class%
-            - %pim_catalog.entity.product_value.class%
+            - '@pim_catalog.doctrine.completeness_query_builder'
